do not use 0-24 delegation for 10.* addresses
This is a stupid hack, but so was the previous state.  It's too
specific to my situation.

Previously, we always set up PTR record assuming that a /24 has been
delegated using CNAMEs, as in RFC 2317.  This is fine for the
situation we normally face, but it's not always the same.

We want to set up 10.* addresses with rdns, and those will not use
RFC 2317 delegation. For those, set up rdns "normally."

What is really needed is a setting that says, for any given IP range,
whether and how to make PTR records.  Later!